System.MIDI.LaunchpadALSA

Description

Interface to the novation Launchpad through ALSA.

Unfortunately, the reset, turn on all LED and double buffering keys etcetera seems to require ugly hacks and occasionally doesn't work. Fortunately, this means that the interface is really simple and it should be possible to get up and running very quickly.

Occasionally when exiting and resuming some program using the launchpad we don't receive any events. A very simple workaround is that, depending on what caused the problem the round buttons to the right or up work. Pushing vol and mixer, for example, should therefore always fix that. Sometimes waiting for a few seconds seems to work also. I suspect this is more annoying while debugging than it will be in use.

Documentation

data Intensity

Intensity of a color, 4 possible values.

Constructors

Off

Low

Med

High

data Color

A color, in intensity of red and green.

Constructors

RG Intensity Intensity

data KeyEvent

A key that was pressed on the launchpad. x, y is like in grid/ungrid, but top buttons have x of 9.

Constructors

Down Word8 Word8

Up Word8 Word8

type App = LaunchpadConnection -> IO ()

Type that withLaunchpad accepts, an IO action that also uses the created handle and connection.

grid :: Word8 -> Word8 -> Word8

Key for the grid where x,y in 0..7, corresponding to x,y from top-left corner. If x is 8 the y corresponds to the side buttons instead. No available mapping for the top buttons afaik.

ungrid :: Word8 -> (Word8, Word8)

Gets the x and y positions from a key.

side :: Word8 -> Word8

Key for side button.

makeData

Arguments

:: Color
Address color from Color

-> Word8
Key made by grid or side

-> Data
Ready for sending by sendData

Create a sendable piece of data.

sendData :: (T DuplexMode, T) -> Data -> IO ()

Send data on connection.

getKey :: (T DuplexMode, T) -> IO KeyEvent

Wait for a keypress and then return it.

listClients :: IO ()

List all the clients for debug purposes (From example of ALSA)

findLaunchpad :: T DuplexMode -> IO (Maybe T)

Finds the launchpad if it exists, by our client handle h

withLaunchpad :: App -> IO ()

Finds the launchpad, and executes f with a client handle and connection. Could be used for one-shots, but it recreates a client and gets the connection with each call. Compose your main to use the given parameters instead, in order to keep the connection open.