System.MIDI.LaunchpadALSA.Extra

Description

Samples and other extras for LaunchpadALSA

You will want to import qualified or otherwise only what you want since main is defined, amongst other things.

Extras & Examples

resetAndLightOnKey :: App

Sample App that turns on all the lights in a pretty pattern on the first event, resets all the lights on the second event (key up most likely), and then quits.

keepLighting :: App

Sample App that keeps lighting the LEDs on keypresses.

lightPressed :: App

Sample App that lights up the key that is pressed (except top buttons since they're unaddressable). Pressing the mixer button exits.

main :: IO ()

Sample main with above App.

allGrid :: [Word8]

Keys for all the grid buttons.

allSide :: [Word8]

Keys for all the side buttons.

allColors :: [Color]

List of all the possible color values (16).

nicePattern :: [Data]

A nice circular pattern.

drawNicePattern :: App

App that draws colors from the centre of the grid.

reset :: App

App that resets all the LEDs.

green :: Color

Basic color green.

red :: Color

Basic color red.

off :: Color

Color value for off.